+++ example/lib/manual_map/dungeon_map.dart
@@ -9,25 +9,25 @@ import 'package:example/shared/enemy/goblin.dart';
 
 class DungeonMap {
   static double tileSize = 45;
-  static final String wallBottom = 'tile/wall_bottom.png';
-  static final String wall = 'tile/wall.png';
-  static final String wallTop = 'tile/wall_top.png';
-  static final String wallLeft = 'tile/wall_left.png';
-  static final String wallBottomLeft = 'tile/wall_bottom_left.png';
-  static final String wallRight = 'tile/wall_right.png';
-  static final String floor_1 = 'tile/floor_1.png';
-  static final String floor_2 = 'tile/floor_2.png';
-  static final String floor_3 = 'tile/floor_3.png';
-  static final String floor_4 = 'tile/floor_4.png';
+  static const String wallBottom = 'tile/wall_bottom.png';
+  static const String wall = 'tile/wall.png';
+  static const String wallTop = 'tile/wall_top.png';
+  static const String wallLeft = 'tile/wall_left.png';
+  static const String wallBottomLeft = 'tile/wall_bottom_left.png';
+  static const String wallRight = 'tile/wall_right.png';
+  static const String floor_1 = 'tile/floor_1.png';
+  static const String floor_2 = 'tile/floor_2.png';
+  static const String floor_3 = 'tile/floor_3.png';
+  static const String floor_4 = 'tile/floor_4.png';
 
   static WorldMap map() {
     List<TileModel> tileList = [];
     List.generate(35, (indexRow) {
-      List.generate(70, (indexColumm) {
-        if (indexRow == 3 && indexColumm > 2 && indexColumm < 30) {
+      List.generate(70, (indexColumn) {
+        if (indexRow == 3 && indexColumn > 2 && indexColumn < 30) {
           tileList.add(TileModel(
             sprite: TileModelSprite(path: wallBottom),
-            x: indexColumm.toDouble(),
+            x: indexColumn.toDouble(),
             y: indexRow.toDouble(),
             collisions: [
               CollisionArea.rectangle(size: Vector2(tileSize, tileSize))
@@ -37,10 +37,10 @@ class DungeonMap {
           ));
           return;
         }
-        if (indexRow == 4 && indexColumm > 2 && indexColumm < 30) {
+        if (indexRow == 4 && indexColumn > 2 && indexColumn < 30) {
           tileList.add(TileModel(
             sprite: TileModelSprite(path: wall),
-            x: indexColumm.toDouble(),
+            x: indexColumn.toDouble(),
             y: indexRow.toDouble(),
             collisions: [
               CollisionArea.rectangle(size: Vector2(tileSize, tileSize))
@@ -51,10 +51,10 @@ class DungeonMap {
           return;
         }
 
-        if (indexRow == 9 && indexColumm > 2 && indexColumm < 30) {
+        if (indexRow == 9 && indexColumn > 2 && indexColumn < 30) {
           tileList.add(TileModel(
             sprite: TileModelSprite(path: wallTop),
-            x: indexColumm.toDouble(),
+            x: indexColumn.toDouble(),
             y: indexRow.toDouble(),
             collisions: [
               CollisionArea.rectangle(size: Vector2(tileSize, tileSize))
@@ -67,12 +67,12 @@ class DungeonMap {
 
         if (indexRow > 4 &&
             indexRow < 9 &&
-            indexColumm > 2 &&
-            indexColumm < 30) {
+            indexColumn > 2 &&
+            indexColumn < 30) {
           tileList.add(
             TileModel(
               sprite: TileModelSprite(path: randomFloor()),
-              x: indexColumm.toDouble(),
+              x: indexColumn.toDouble(),
               y: indexRow.toDouble(),
               width: tileSize,
               height: tileSize,
@@ -81,10 +81,10 @@ class DungeonMap {
           return;
         }
 
-        if (indexRow > 3 && indexRow < 9 && indexColumm == 2) {
+        if (indexRow > 3 && indexRow < 9 && indexColumn == 2) {
           tileList.add(TileModel(
             sprite: TileModelSprite(path: wallLeft),
-            x: indexColumm.toDouble(),
+            x: indexColumn.toDouble(),
             y: indexRow.toDouble(),
             collisions: [
               CollisionArea.rectangle(size: Vector2(tileSize, tileSize))
@@ -93,10 +93,10 @@ class DungeonMap {
             height: tileSize,
           ));
         }
-        if (indexRow == 9 && indexColumm == 2) {
+        if (indexRow == 9 && indexColumn == 2) {
           tileList.add(TileModel(
             sprite: TileModelSprite(path: wallBottomLeft),
-            x: indexColumm.toDouble(),
+            x: indexColumn.toDouble(),
             y: indexRow.toDouble(),
             collisions: [
               CollisionArea.rectangle(size: Vector2(tileSize, tileSize))
@@ -106,10 +106,10 @@ class DungeonMap {
           ));
         }
 
-        if (indexRow > 3 && indexRow < 9 && indexColumm == 30) {
+        if (indexRow > 3 && indexRow < 9 && indexColumn == 30) {
           tileList.add(TileModel(
             sprite: TileModelSprite(path: wallRight),
-            x: indexColumm.toDouble(),
+            x: indexColumn.toDouble(),
             y: indexRow.toDouble(),
             collisions: [
               CollisionArea.rectangle(size: Vector2(tileSize, tileSize))