+++ example/lib/shared/interface/bar_life_component.dart
@@ -18,12 +18,15 @@ class BarLifeInterface extends InterfaceComponent
         );
 
   @override
-  void render(Canvas c) {
-    super.render(c);
+  void render(Canvas canvas) {
+    super.render(canvas);
     try {
-      _drawLife(c);
-      _drawStamina(c);
-    } catch (e) {}
+      _drawLife(canvas);
+      _drawStamina(canvas);
+    } catch (e) {
+      // ignore: avoid_print
+      print(e);
+    }
   }
 
   void _drawLife(Canvas canvas) {