+++ lib/collision/collision_config.dart
@@ -36,7 +36,6 @@ class CollisionConfig {
   void updatePosition(Vector2 position) {
     if (collisions.isNotEmpty && position != _lastPosition) {
       _lastPosition = position.clone();
-      collisions.first.updatePosition(position);
       Rect? newRect;
       for (var element in collisions) {
         element.updatePosition(position);